fix tests & refactor
