+++ posthog/api/authentication.py
@@ -6,7 +6,7 @@ from django.contrib.auth import views as auth_views
 from django.contrib.auth.password_validation import validate_password
 from django.contrib.auth.tokens import default_token_generator
 from django.core.exceptions import ValidationError
-from django.http import JsonResponse
+from django.http import HttpRequest, HttpResponse, JsonResponse
 from django.shortcuts import redirect
 from django.utils import timezone
 from django.views.decorators.csrf import csrf_protect
@@ -14,10 +14,12 @@ from loginas.utils import is_impersonated_session, restore_original_login
 from rest_framework import mixins, permissions, serializers, status, viewsets
 from rest_framework.request import Request
 from rest_framework.response import Response
+from social_django.views import auth
 
 from posthog.email import EmailMessage, is_email_available
 from posthog.event_usage import report_user_logged_in, report_user_password_reset
 from posthog.models import OrganizationDomain, User
+from posthog.utils import get_instance_available_sso_providers
 
 
 @csrf_protect
@@ -47,6 +49,20 @@ def axes_locked_out(*args, **kwargs):
         },
         status=status.HTTP_403_FORBIDDEN,
     )
+
+
+def sso_login(request: HttpRequest, backend: str) -> HttpResponse:
+    sso_providers = get_instance_available_sso_providers()
+    # because SAML is configured at the domain-level, we have to assume it's enabled for someone in the instance
+    sso_providers["saml"] = settings.EE_AVAILABLE
+
+    if backend not in sso_providers:
+        return redirect(f"/login?error_code=invalid_sso_provider")
+
+    if not sso_providers[backend]:
+        return redirect(f"/login?error_code=improperly_configured_sso")
+
+    return auth(request, backend)
 
 
 class LoginSerializer(serializers.Serializer):
